Split mercator_utils tests into a separate file
