+++ src/util/test/util.ts
@@ -216,3 +216,10 @@ export function createStyle() {
         layers: []
     } as StyleSpecification;
 }
+
+export function expectToBeCloseToArray(actual: Array<number>, expected: Array<number>, precision?: number) {
+    expect(actual).toHaveLength(expected.length);
+    for (let i = 0; i < expected.length; i++) {
+        expect(actual[i]).toBeCloseTo(expected[i], precision);
+    }
+}
